Update RandomGenerator.py
included loops with validation to ensure that the user inputs at least one keyword and a positive number for the number of results they want to generate. It also uses try-except block to validate the input of number of results if the user inputs something that cannot be converted to int.

It also includes error handling with the requests module, it will print an error message if there is